Separate out shared test sources for couchbase/play-ws to another project.
