+++ instrumentation/couchbase/couchbase-2.6/src/test/groovy/CouchbaseClient26Test.groovy
@@ -16,7 +16,7 @@
 
 import io.opentelemetry.auto.test.asserts.TraceAssert
 
-class CouchbaseClient26Test extends CouchbaseClientTest {
+class CouchbaseClient26Test extends AbstractCouchbaseClientTest {
   @Override
   void assertCouchbaseCall(TraceAssert trace, int index, Object name, String bucketName = null, Object parentSpan = null) {
     CouchbaseSpanUtil.assertCouchbaseCall(trace, index, name, bucketName, parentSpan)
