+++ instrumentation/couchbase/couchbase-2.6/src/test/groovy/springdata/CouchbaseSpringRepository26Test.groovy
@@ -18,7 +18,7 @@ package springdata
 
 import io.opentelemetry.auto.test.asserts.TraceAssert
 
-class CouchbaseSpringRepository26Test extends CouchbaseSpringRepositoryTest {
+class CouchbaseSpringRepository26Test extends AbstractCouchbaseSpringRepositoryTest {
 
   @Override
   void assertCouchbaseCall(TraceAssert trace, int index, Object name, String bucketName = null, Object parentSpan = null) {
