+++ instrumentation/couchbase/couchbase-2.6/src/test/groovy/springdata/CouchbaseSpringTemplate26Test.groovy
@@ -18,7 +18,7 @@ package springdata
 
 import io.opentelemetry.auto.test.asserts.TraceAssert
 
-class CouchbaseSpringTemplate26Test extends CouchbaseSpringTemplateTest {
+class CouchbaseSpringTemplate26Test extends AbstractCouchbaseSpringTemplateTest {
   @Override
   void assertCouchbaseCall(TraceAssert trace, int index, Object name, String bucketName = null, Object parentSpan = null) {
     CouchbaseSpanUtil.assertCouchbaseCall(trace, index, name, bucketName, parentSpan)
