+++ instrumentation/couchbase/testing/src/main/groovy/springdata/CouchbaseConfig.groovy
