+++ instrumentation/couchbase/testing/src/main/groovy/util/AbstractCouchbaseTest.groovy
