+++ instrumentation/play-ws/play-ws-1.0/play-ws-1.0.gradle
@@ -38,7 +38,8 @@ dependencies {
   compileOnly group: 'com.typesafe.play', name: "play-ahc-ws-standalone_$scalaVersion", version: '1.0.2'
 
   implementation project(':instrumentation:play-ws:play-ws-common')
-  testImplementation project(path: ':instrumentation:play-ws:play-ws-common', configuration: 'testArtifacts')
+
+  testImplementation project(':instrumentation:play-ws:testing')
 
   // These are to ensure cross compatibility
   testImplementation project(':instrumentation:netty:netty-4.0')
