+++ settings.gradle
@@ -67,6 +67,7 @@ include ':instrumentation:cassandra:cassandra-4.0'
 include ':instrumentation:cdi-testing'
 include ':instrumentation:couchbase:couchbase-2.0'
 include ':instrumentation:couchbase:couchbase-2.6'
+include ':instrumentation:couchbase:testing'
 include ':instrumentation:dropwizard-views-0.7'
 include ':instrumentation:dropwizard-testing'
 include ':instrumentation:elasticsearch:elasticsearch-rest-common'
@@ -137,6 +138,7 @@ include ':instrumentation:play-ws:play-ws-1.0'
 include ':instrumentation:play-ws:play-ws-2.0'
 include ':instrumentation:play-ws:play-ws-2.1'
 include ':instrumentation:play-ws:play-ws-common'
+include ':instrumentation:play-ws:testing'
 include ':instrumentation:rabbitmq-2.7'
 include ':instrumentation:ratpack-1.4'
 include ':instrumentation:reactor-3.1'
